requirements.txt
streamlit
transformers
torch
torchvision
Pillow
requests
python-dotenv
